+++ spotdl/metadata_search.py
@@ -86,7 +86,7 @@ class MetadataSearch:
             "query": self._on_youtube_and_spotify_for_type_query,
         }
         caller = track_type_mapper[self.track_type]
-        metadata = caller()
+        metadata = caller(self.track)
 
         if not self.lyrics:
             return metadata
@@ -173,16 +173,16 @@ class MetadataSearch:
         return video
 
     def _best_on_youtube_search_for_type_spotify(self, url):
-        spotify_metadata = self._on_spotify_for_type_spotify(self.track)
+        spotify_metadata = self._on_spotify_for_type_spotify(url)
         search_query = spotdl.metadata.format_string(self.yt_search_format, spotify_metadata)
         video = self._best_on_youtube_search_for_type_query(search_query)
         return video
 
-    def _on_youtube_and_spotify_for_type_spotify(self):
+    def _on_youtube_and_spotify_for_type_spotify(self, url):
         logger.debug("Extracting YouTube and Spotify metadata for input Spotify URI.")
-        spotify_metadata = self._on_spotify_for_type_spotify(self.track)
+        spotify_metadata = self._on_spotify_for_type_spotify(url)
         search_query = spotdl.metadata.format_string(self.yt_search_format, spotify_metadata)
-        youtube_video = self._best_on_youtube_search_for_type_spotify(search_query)
+        youtube_video = self._best_on_youtube_search_for_type_query(search_query)
         youtube_metadata = self.providers["youtube"].from_url(youtube_video["url"])
         metadata = spotdl.util.merge_copy(
             youtube_metadata,
@@ -190,9 +190,9 @@ class MetadataSearch:
         )
         return metadata
 
-    def _on_youtube_and_spotify_for_type_youtube(self):
+    def _on_youtube_and_spotify_for_type_youtube(self, url):
         logger.debug("Extracting YouTube and Spotify metadata for input YouTube URL.")
-        youtube_metadata = self._on_youtube_for_type_youtube(self.track)
+        youtube_metadata = self._on_youtube_for_type_youtube(url)
         search_query = spotdl.metadata.format_string("{track-name}", youtube_metadata)
         spotify_metadata = self._on_spotify_for_type_query(search_query)
         metadata = spotdl.util.merge_copy(
@@ -201,28 +201,25 @@ class MetadataSearch:
         )
         return metadata
 
-    def _on_youtube_and_spotify_for_type_query(self):
+    def _on_youtube_and_spotify_for_type_query(self, query):
         logger.debug("Extracting YouTube and Spotify metadata for input track query.")
-        search_query = self.track
         # Make use of threads here to search on both YouTube & Spotify
         # at the same time.
         spotify_metadata = spotdl.util.ThreadWithReturnValue(
             target=self._on_spotify_for_type_query,
-            args=(search_query,)
+            args=(query,)
         )
         spotify_metadata.start()
-        youtube_metadata = self._on_youtube_for_type_query(search_query)
+        youtube_metadata = self._on_youtube_for_type_query(query)
         metadata = spotdl.util.merge_copy(
             youtube_metadata,
             spotify_metadata.join()
         )
         return metadata
 
-    def _on_youtube_for_type_spotify(self):
+    def _on_youtube_for_type_spotify(self, url):
         logger.debug("Extracting YouTube metadata for input Spotify URI.")
-        spotify_metadata = self._on_spotify_for_type_spotify(self.track)
-        search_query = spotdl.metadata.format_string(self.yt_search_format, spotify_metadata)
-        youtube_video = self._best_on_youtube_search_for_type_spotify(search_query)
+        youtube_video = self._best_on_youtube_search_for_type_spotify(url)
         youtube_metadata = self.providers["youtube"].from_url(youtube_video["url"])
         return youtube_metadata